GCP Scheduler extension should use shared handler
