+++ package.json
@@ -41,5 +41,8 @@
       "pre-commit": "npm run lint",
       "pre-push": "npm run lint && npm test"
     }
+  },
+  "dependencies": {
+    "gcp-jupyterlab-shared": "^1.0.0"
   }
 }
